requirements.txt
apturl==0.5.2
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
click==8.0.3
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-jsonapi==5.0.0
entrypoints==0.3
Flask==2.0.2
glob2==0.7
h11==0.13.0
httplib2==0.14.0
idna==2.8
inflection==0.5.1
itsdangerous==2.0.1
Jinja2==3.0.3
junit-xml==1.8
kafka==1.3.5
kafka-python==2.0.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
MarkupSafe==2.0.1
mrjob==0.7.4
mutmut==2.2.0
mysqlclient==2.1.0
netifaces==0.10.4
numpy==1.21.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.4.60
outcome==1.1.0
pandas==1.4.1
parso==0.8.3
pexpect==4.6.0
Pillow==7.0.0
pip==20.0.2
pony==0.7.14
protobuf==3.6.1
py4j==0.10.9.2
pycairo==1.16.2
pycodestyle==2.8.0
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==4.0.1
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyRFC3339==1.1
PySocks==1.7.1
pyspark==3.2.0
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2021.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selenium==4.1.2
setuptools==45.2.0
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
systemd-python==234
toml==0.10.2
trio==0.20.0
trio-websocket==0.9.2
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
wadllib==1.3.3
Werkzeug==2.0.2
wheel==0.34.2
wsproto==1.1.0
xkit==0.0.0
